pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/aitelegrambot"]
[project]
name = "aitelegrambot"
version = "0.5.2"
dependencies = [
  "ollama",
  "python-telegram-bot",
  "python-dotenv",
]
authors = [
  {name = "tusharhero", email = "tusharhero@sdf.org"},
  {name = "alokosx", email = "alokosx@gmail.com"},
]
description = "aitelegrambot is a Telegram bot that uses the Ollama backend to run the LLM rationalAI (by default)."
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]
scripts = { aitelegrambot = "aitelegrambot.__main__:main" }
